save development state
