+++ core/src/pkg.rs
@@ -38,4 +38,3 @@ impl<'a> LodPkg<'a> {
         }
     }
 }
-
